backend/pyproject.toml
[tool.poetry]
name = "toody-backend"
version = "0.1.0"
description = ""
authors = ["bartolkaruza <bartol.k@gmail.com>"]
readme = "README.md"
packages = [{include = "toody_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
numpy = "^1.25.0"
gunicorn = "^20.1.0"
faiss-cpu = "^1.7.4"
scikit-learn = "^1.2.2"
transformers = "^4.30.2"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
